Apply 100_Codechange_Move_landscape_functions_to_landscape_files.diff
I checked the code for FixSlopes() and FlatEmptyWorld() and it hasn't
changed since this patch was created, so this is just moving the current
code and not throwing away any recent bug fixes.
