+++ source.list
@@ -210,7 +210,6 @@ group_gui.h
 group_type.h
 gui.h
 guitimer_func.h
-heightmap.h
 heightmap_base.h
 heightmap_layer_base.h
 heightmap_layer_type.h
