+++ src/heightmap.cpp
@@ -10,7 +10,6 @@
 /** @file heightmap.cpp Creating of maps from heightmaps. */
 
 #include "stdafx.h"
-#include "heightmap.h"
 #include "heightmap_type.h"
 #include "heightmap_base.h"
 #include "clear_map.h"
@@ -289,66 +288,6 @@ static bool ReadHeightmapBMP(const char *filename, uint *x, uint *y, byte **map)
 }
 
 /**
- * This function takes care of the fact that land in OpenTTD can never differ
- * more than 1 in height
- */
-void FixSlopes()
-{
-	uint width, height;
-	int row, col;
-	byte current_tile;
-
-	/* Adjust height difference to maximum one horizontal/vertical change. */
-	width   = MapSizeX();
-	height  = MapSizeY();
-
-	/* Top and left edge */
-	for (row = 0; (uint)row < height; row++) {
-		for (col = 0; (uint)col < width; col++) {
-			current_tile = MAX_TILE_HEIGHT;
-			if (col != 0) {
-				/* Find lowest tile; either the top or left one */
-				current_tile = TileHeight(TileXY(col - 1, row)); // top edge
-			}
-			if (row != 0) {
-				if (TileHeight(TileXY(col, row - 1)) < current_tile) {
-					current_tile = TileHeight(TileXY(col, row - 1)); // left edge
-				}
-			}
-
-			/* Does the height differ more than one? */
-			if (TileHeight(TileXY(col, row)) >= (uint)current_tile + 2) {
-				/* Then change the height to be no more than one */
-				SetTileHeight(TileXY(col, row), current_tile + 1);
-			}
-		}
-	}
-
-	/* Bottom and right edge */
-	for (row = height - 1; row >= 0; row--) {
-		for (col = width - 1; col >= 0; col--) {
-			current_tile = MAX_TILE_HEIGHT;
-			if ((uint)col != width - 1) {
-				/* Find lowest tile; either the bottom and right one */
-				current_tile = TileHeight(TileXY(col + 1, row)); // bottom edge
-			}
-
-			if ((uint)row != height - 1) {
-				if (TileHeight(TileXY(col, row + 1)) < current_tile) {
-					current_tile = TileHeight(TileXY(col, row + 1)); // right edge
-				}
-			}
-
-			/* Does the height differ more than one? */
-			if (TileHeight(TileXY(col, row)) >= (uint)current_tile + 2) {
-				/* Then change the height to be no more than one */
-				SetTileHeight(TileXY(col, row), current_tile + 1);
-			}
-		}
-	}
-}
-
-/**
  * Reads the heightmap with the correct file reader.
  * @param dft Type of image file.
  * @param filename Name of the file to load.
@@ -371,23 +310,6 @@ static bool ReadHeightMap(DetailedFileType dft, const char *filename, uint *x, u
 		case DFT_HEIGHTMAP_BMP:
 			return ReadHeightmapBMP(filename, x, y, map);
 	}
-}
-
-/**
- * Make an empty world where all tiles are of height 'tile_height'.
- * @param tile_height of the desired new empty world
- */
-void FlatEmptyWorld(byte tile_height)
-{
-	int edge_distance = _settings_game.construction.freeform_edges ? 0 : 2;
-	for (uint row = edge_distance; row < MapSizeY() - edge_distance; row++) {
-		for (uint col = edge_distance; col < MapSizeX() - edge_distance; col++) {
-			SetTileHeight(TileXY(col, row), tile_height);
-		}
-	}
-
-	FixSlopes();
-	MarkWholeScreenDirty();
 }
 
 /**
