--- src/heightmap.h
@@ -1,20 +1,0 @@
-/* $Id$ */
-
-/*
- * This file is part of OpenTTD.
- * OpenTTD is free software; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, version 2.
- * OpenTTD is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU General Public License for more details. You should have received a copy of the GNU General Public License along with OpenTTD. If not, see <http://www.gnu.org/licenses/>.
- */
-
-/** @file heightmap.h Functions related to creating heightmaps from files. */
-
-#ifndef HEIGHTMAP_H
-#define HEIGHTMAP_H
-
-#include "fileio_type.h"
-
-void FlatEmptyWorld(byte tile_height);
-void FixSlopes();
-
-#endif /* HEIGHTMAP_H */
